feat: extract checklist subcomponents
